requirements-dev.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
coverage==5.2.1
flake8==3.8.3
importlib-metadata==1.7.0
iniconfig==1.0.1
lark-parser==0.9.0
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.0.1
pytest-black==0.3.10
pytest-cov==2.10.0
pytest-flake8==1.0.6
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
zipp==3.1.0
